Fix code review feedback: DoS vulnerability, type consistency, logging
## Critical Fixes

1. **Fix validate_actions DoS vulnerability** (crates/forge-core/src/systems.rs)
   - Changed loop to iterate over agents instead of actions
   - Use actions.get(i).unwrap_or(&Action::Noop) for safe access
   - Always returns exactly one action per agent
   - Prevents out-of-bounds panic when fewer actions than agents provided
   - Updated tests to reflect new behavior

2. **Change constants from f32 to f64** (crates/forge-types/src/constants.rs)
   - Changed all BIOME_* constants from f32 to f64
   - Changed TERRAIN_NOISE_OCTAVES_MULTIPLIER from f32 to f64
   - Aligns with TERRAIN_NOISE_PERSISTENCE (already f64)
   - Eliminates unnecessary casting in worldgen code
   - Removed all as f64 casts in biome.rs and terrain.rs

3. **Fix logging consistency** (crates/forge-worldgen/src/biome.rs)
   - Changed tracing::trace!(...) to trace!(...)
   - Consistent with imported trace macro
   - Unified logging style

4. **Optimize object lookup** (crates/forge-core/src/systems.rs)
   - Pre-compute HashMap<id, &Object> for O(1) lookups
   - Replaces linear search inside nested loops
   - Improves performance from O(agents × positions × objects) to O(agents × positions)

5. **Updated tests**
   - test_validate_actions_empty now expects 1 Noop action (1 agent)
   - test_validate_actions_excess_actions expects 1 action (1 agent, ignores extras)
   - Added test_validate_actions_insufficient_actions for safety case

## Verification
✅ All 569+ tests passing
✅ cargo clippy with zero warnings
✅ cargo fmt compliant
✅ No breaking changes

+++ crates/forge-types/src/constants.rs
@@ -105,29 +105,29 @@ pub const FIXED_POINT_SHIFT: u32 = 16;
 pub const FIXED_POINT_ONE: i32 = 1 << FIXED_POINT_SHIFT; // 65536
 
 /// Base water level threshold for biome classification (before scale adjustment).
-pub const BIOME_WATER_LEVEL_BASE: f32 = 0.35;
+pub const BIOME_WATER_LEVEL_BASE: f64 = 0.35;
 /// Water level adjustment rate per biome scale unit.
-pub const BIOME_WATER_LEVEL_SCALE_MULTIPLIER: f32 = 0.3;
+pub const BIOME_WATER_LEVEL_SCALE_MULTIPLIER: f64 = 0.3;
 /// Minimum water level threshold.
-pub const BIOME_WATER_LEVEL_MIN: f32 = 0.10;
+pub const BIOME_WATER_LEVEL_MIN: f64 = 0.10;
 /// Maximum water level threshold.
-pub const BIOME_WATER_LEVEL_MAX: f32 = 0.50;
+pub const BIOME_WATER_LEVEL_MAX: f64 = 0.50;
 /// Base mountain level threshold for biome classification.
-pub const BIOME_MOUNTAIN_LEVEL_BASE: f32 = 0.72;
+pub const BIOME_MOUNTAIN_LEVEL_BASE: f64 = 0.72;
 /// Mountain level adjustment rate per biome scale unit.
-pub const BIOME_MOUNTAIN_LEVEL_SCALE_MULTIPLIER: f32 = 0.3;
+pub const BIOME_MOUNTAIN_LEVEL_SCALE_MULTIPLIER: f64 = 0.3;
 /// Minimum mountain level threshold.
-pub const BIOME_MOUNTAIN_LEVEL_MIN: f32 = 0.60;
+pub const BIOME_MOUNTAIN_LEVEL_MIN: f64 = 0.60;
 /// Maximum mountain level threshold.
-pub const BIOME_MOUNTAIN_LEVEL_MAX: f32 = 0.90;
+pub const BIOME_MOUNTAIN_LEVEL_MAX: f64 = 0.90;
 /// Sand level offset from water level.
-pub const BIOME_SAND_LEVEL_OFFSET: f32 = 0.05;
+pub const BIOME_SAND_LEVEL_OFFSET: f64 = 0.05;
 /// Moisture threshold for forest biome.
-pub const BIOME_FOREST_MOISTURE_THRESHOLD: f32 = 0.45;
+pub const BIOME_FOREST_MOISTURE_THRESHOLD: f64 = 0.45;
 /// Moisture threshold for desert biome.
-pub const BIOME_DESERT_MOISTURE_THRESHOLD: f32 = 0.25;
+pub const BIOME_DESERT_MOISTURE_THRESHOLD: f64 = 0.25;
 /// Number of octaves for terrain noise generation (before scale adjustment).
-pub const TERRAIN_NOISE_OCTAVES_MULTIPLIER: f32 = 40.0;
+pub const TERRAIN_NOISE_OCTAVES_MULTIPLIER: f64 = 40.0;
 /// Minimum octaves for terrain noise.
 pub const TERRAIN_NOISE_OCTAVES_MIN: u32 = 2;
 /// Maximum octaves for terrain noise.
